+++ engine/com/kontechs/kje/Rectangle.java
@@ -1,9 +1,9 @@
 package com.kontechs.kje;
 
 public class Rectangle {
-	public int x, y, width, height;
+	public double x, y, width, height;
 	
-	public Rectangle(int x, int y, int width, int height) {
+	public Rectangle(double x, double y, double width, double height) {
 		this.x = x;
 		this.y = y;
 		this.width = width;
